Added worded version (e.g. alpha 3) to versioning system. Changed version to 0.2. Moved export button to main window, added bottom panel there. Auto-increment build+revision version.
